+++ drive/segment_analysis/pre_shared_segments_analysis_scripts/shared_segment_detection/gathering_pairs/find_IBD_phenotype.py
@@ -3,7 +3,7 @@ import pandas as pd
 
 import multiprocessing as mp
 from functools import partial
-
+from typing import Dict, Optional
 from .collect_shared_segments import generate_parameters, build_unique_id_dict, create_ibd_arrays, gather_pairs
 import utility_scripts
 
@@ -89,7 +89,7 @@ def get_ibd_file(ibd_list: list, chr_num) -> str:
 
     return [file for file in ibd_list if chr_num in file][0]
 
-def collect_IBD_segments(carrier_list: list, ibd_program:str, min_CM: str, ibd_file_list: list, output_path: str, gene_dict: dict, que_object, key: str):
+def collect_IBD_segments(carrier_list: list, ibd_program:str, min_CM: str, ibd_file_list: list, output_path: str, gene_dict: dict, pair_info_dict: Dict[str, Dict], que_object, key: str):
 
     gene_info: dict = gene_dict[key]
     
@@ -104,23 +104,27 @@ def collect_IBD_segments(carrier_list: list, ibd_program:str, min_CM: str, ibd_f
 
     IBDdata, IBDindex = create_ibd_arrays()
 
-    gather_pairs(IBDdata, IBDindex, parameter_dict, ibd_file, uniqID, min_CM, que_object, output_path, ibd_program, gene_start=gene_info["start"], gene_end=gene_info["end"], gene_name=key) 
+    gather_pairs(IBDdata, IBDindex, parameter_dict, ibd_file, uniqID, min_CM, que_object, output_path, ibd_program, pair_info_dict, gene_start=gene_info["start"], gene_end=gene_info["end"], gene_name=key) 
 
-def run_parallel(gene_info_dict: dict, ibd_file_list: list,THREADS: int, min_CM: str, ibd_program: str, output: str, carrier_list: list):
+def run_parallel(gene_info_dict: dict, ibd_file_list: list,THREADS: int, min_CM: str, ibd_program: str, output: str, carrier_list: list, ) -> Dict:
     """function to run through the genes in parallel"""
+    
 
     manager = mp.Manager()
 
     que = manager.Queue()
 
     pool = mp.Pool(int(THREADS))
+
+    pair_info_dict: Dict[str, Dict] = manager.dict()
+
     header:str = "gene\n"
 
     watcher = pool.apply_async(
             utility_scripts.listener,
             (que, "".join([output, "gene_target_failed.txt"]), header))
 
-    func = partial(collect_IBD_segments, carrier_list, ibd_program, min_CM, ibd_file_list, output, gene_info_dict, que)
+    func = partial(collect_IBD_segments, carrier_list, ibd_program, min_CM, ibd_file_list, output, gene_info_dict, pair_info_dict, que)
 
     pool.map(func, list(gene_info_dict.keys()))
 
@@ -130,8 +134,10 @@ def run_parallel(gene_info_dict: dict, ibd_file_list: list,THREADS: int, min_CM:
 
     pool.join()
 
+    return pair_info_dict
 
-def gather_shared_segments(ibd_file_list: list, pheno_gmap_df:pd.DataFrame, phenotype_carriers_df: pd.DataFrame, output_path: str, ibd_program: str, min_CM: str, ibd_suffix: str, THREADS):
+
+def gather_shared_segments(ibd_file_list: list, pheno_gmap_df:pd.DataFrame, phenotype_carriers_df: pd.DataFrame, output_path: str, ibd_program: str, min_CM: str, ibd_suffix: str, THREADS, pair_info_dict) -> Dict[str, Dict]:
     """Function to get the shared segments for each pair within a gene of interest
     Parameters
     __________
@@ -152,6 +158,9 @@ def gather_shared_segments(ibd_file_list: list, pheno_gmap_df:pd.DataFrame, phen
     
     ibd_program : str
         This is the ibd program used to get the shared segment data. Should be either ilash or hapibd"""
+
+    # adding the ibd program as a key to the dictionary
+    pair_info_dict.setdefault(ibd_program, {})
     
     # checking to make sure the output directory subdirectory "collected_pairs" exists
     ibd_output_path: str = utility_scripts.check_dir(output_path, "collected_pairs/")
@@ -165,5 +174,10 @@ def gather_shared_segments(ibd_file_list: list, pheno_gmap_df:pd.DataFrame, phen
     # need to generate a dictionary of all chromosomes, with their start and end point
     gene_dict: dict = gather_gene_info(pheno_gmap_df)
 
-    run_parallel(gene_dict, ibd_file_list, THREADS, min_CM, ibd_program, ibd_output_path, carrier_list)
+
+    pairs_dict: Dict[str, Dict] = run_parallel(gene_dict, ibd_file_list, THREADS, min_CM, ibd_program, ibd_output_path, carrier_list)
+
+    for key, value in pairs_dict.items():
+
+        pair_info_dict[ibd_program].setdefault(key, value)