+++ drive/segment_analysis/pre_shared_segments_analysis_scripts/shared_segment_detection/gathering_pairs/gather_ibd_info.py
@@ -145,9 +145,6 @@ def collect_files(parameter_dict: dict, carrier_file: str) -> dict:
     # getting the segment file list
     
     segment_file_list: List[str] = utility_scripts.get_file_list(ibd_files, file_suffix)
-    
-    # print(carrier_file)
-    # chr_var_file_list: List[str] = utility_scripts.get_file_list(carrier_file, "*.single_variant_carrier.csv")
    
 
     # getting the list of map files
@@ -230,7 +227,7 @@ def filter_no_carriers(var_iid_dict: dict, output_path: str, chr_num: str) -> di
 
 #TODO: refactor to make this function testable
 # This function is not testable at the moment
-def iterate_file_dict(file_dict: dict, output: str, threads: str, ibd_program: str, min_CM: str):
+def iterate_file_dict(file_dict: dict, output: str, threads: str, ibd_program: str, min_CM: str, pair_info_dict: Dict[str, Dict]) -> Dict[str, Dict]:
     """Function will iterate through the file dictionary which has paired the chromosome number with the appropriate files 
     Parameters
     __________
@@ -245,6 +242,11 @@ def iterate_file_dict(file_dict: dict, output: str, threads: str, ibd_program: s
     output : str
         string that list directory to output files at
     """
+
+    
+
+    pair_info_dict.setdefault(ibd_program, {})
+
     # Iterating through the chromosomes that have a value
     for key in file_dict:
         
@@ -275,6 +277,7 @@ def iterate_file_dict(file_dict: dict, output: str, threads: str, ibd_program: s
                 var_info_dict = create_var_info_dict(var_info_dict, var_iid_dict, variant, bp)
 
             # need to fix this part for the new function
+            #TODO: Need to get rid of the decorator
             parallel_runner: object = utility_scripts.Segment_Parallel_Runner(
                 int(threads), output, ibd_program, min_CM, var_info_dict,
                 ibd_file)
@@ -282,26 +285,35 @@ def iterate_file_dict(file_dict: dict, output: str, threads: str, ibd_program: s
             error_filename: str = "nopairs-identified.txt"
             header_str: str = "variant_id"
 
-            parallel_runner.run_segments_parallel(error_filename, gather_shared_segments,
-                                                  header_str)
+            pairs_dict: Dict[str, Dict] = parallel_runner.run_segments_parallel(gather_shared_segments, 
+                    error_filename, header_str)
+            
+            
+            for key, value in pairs_dict.items():
+
+                pair_info_dict[ibd_program].setdefault(key, value)
+
+
+
 
 # TODO: rename function
-def gather_shared_segments(segment_file: str, output_path: str, ibd_format: str,
-             min_CM: str, var_info_dict: list, que_object, variant):
+def gather_shared_segments(variant: str, segment_file: str, output_path: str, ibd_format: str,
+             min_CM: str, var_info_dict: list, pair_info_dict: Dict[str, Dict], que_object):
 
     output_path: str = utility_scripts.check_dir(output_path, "collected_pairs/")
+
     variant_position: int = int(var_info_dict[variant]["base_pos"])
 
     print(f"running the variant {variant}")
 
-    carrier_list: list = var_info_dict[variant]["iid_list"]
+    carrier_list: List = var_info_dict[variant]["iid_list"]
 
-    parameter_dict: dict = generate_parameters(ibd_format)
+    parameter_dict: Dict = generate_parameters(ibd_format)
     
-    uniqID: dict = build_unique_id_dict(carrier_list)
+    uniqID: Dict = build_unique_id_dict(carrier_list)
 
     IBDdata, IBDindex = create_ibd_arrays()
 
-    gather_pairs(IBDdata, IBDindex, parameter_dict, segment_file, uniqID, min_CM, que_object, output_path, ibd_format, var_position=variant_position, variant_name=variant) 
+    gather_pairs(IBDdata, IBDindex, parameter_dict, segment_file, uniqID, min_CM, que_object, output_path, ibd_format, pair_info_dict, var_position=variant_position, variant_name=variant)